+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 - Click [here](https://r35007.github.io/Mock-Server/) for mock server documentation.
 - Click [here](https://github.com/R35007/Mock-Server/blob/main/CHANGELOG.md) to see more changes reg the `@r35007/mock-server` package
 
+## v20.1.0
+
+- Updated - `@35007/mock-server@18.1.1` node package.
+- Added - `MockServer: Start Server in Terminal` - THis command starts the mock server in terminal
+
 ## v20.0.0
 
 - Updated - `@35007/mock-server@18.1.0` node package.